# backend/docs/server/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Server for CI/CD application developed using Go (Gin, Gorm).",
        "title": "CI/CD Application - Server API",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/api",
    "paths": {
        "/build_steps": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "build steps"
                ],
                "summary": "Get all build steps",
                "operationId": "all-build-steps",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Order by field",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "description": "Filter",
                        "name": "build_step",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/model.BuildStepCore"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of build steps",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.BuildStepShort"
                            }
                        }
                    },
                    "400": {
                        "description": "Error in request",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    },
                    "404": {
                        "description": "No records found",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    }
                }
            }
        },
        "/build_steps/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "build steps"
                ],
                "summary": "Get the single build step",
                "operationId": "single-build-step",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Build step ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Requested build step",
                        "schema": {
                            "$ref": "#/definitions/model.BuildStep"
                        }
                    },
                    "400": {
                        "description": "Error in params",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    },
                    "404": {
                        "description": "No record found",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    }
                }
            }
        },
        "/builds": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "builds"
                ],
                "summary": "Get all builds",
                "operationId": "all-builds",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Order by field",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "description": "Filter",
                        "name": "build",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/model.BuildCore"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of builds",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.BuildShort"
                            }
                        }
                    },
                    "400": {
                        "description": "Error in request",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    },
                    "404": {
                        "description": "No records found",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "builds"
                ],
                "summary": "Create new build",
                "operationId": "create-build",
                "parameters": [
                    {
                        "description": "New build entry",
                        "name": "build",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.BuildCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Newly created build",
                        "schema": {
                            "$ref": "#/definitions/model.Build"
                        }
                    },
                    "400": {
                        "description": "Error in params",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    },
                    "501": {
                        "description": "Endpoint not implemented",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    }
                }
            }
        },
        "/builds/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "builds"
                ],
                "summary": "Get the single build",
                "operationId": "single-build",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Build ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Requested build",
                        "schema": {
                            "$ref": "#/definitions/model.Build"
                        }
                    },
                    "400": {
                        "description": "Error in params",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    },
                    "404": {
                        "description": "No record found",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "builds"
                ],
                "summary": "Update build",
                "operationId": "update-build",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Build ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated build entry",
                        "name": "build",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.BuildCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated build",
                        "schema": {
                            "$ref": "#/definitions/model.Build"
                        }
                    },
                    "400": {
                        "description": "Error in params",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    },
                    "501": {
                        "description": "Endpoint not implemented",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.Build": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "number": {
                    "type": "integer"
                },
                "pipeline_id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.BuildStatus"
                },
                "steps": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.BuildStep"
                    }
                },
                "tags": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "worker_id": {
                    "type": "string"
                }
            }
        },
        "model.BuildCore": {
            "type": "object",
            "properties": {
                "number": {
                    "type": "integer"
                },
                "pipeline_id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.BuildStatus"
                },
                "tags": {
                    "type": "string"
                },
                "worker_id": {
                    "type": "string"
                }
            }
        },
        "model.BuildCreate": {
            "type": "object",
            "properties": {
                "number": {
                    "type": "integer"
                },
                "pipeline_id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.BuildStatus"
                },
                "tags": {
                    "type": "string"
                },
                "worker_id": {
                    "type": "string"
                }
            }
        },
        "model.BuildLog": {
            "type": "object",
            "properties": {
                "build_step_id": {
                    "type": "string"
                },
                "command": {
                    "type": "string"
                },
                "output": {
                    "type": "string"
                }
            }
        },
        "model.BuildShort": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "number": {
                    "type": "integer"
                },
                "pipeline_id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.BuildStatus"
                },
                "steps": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.BuildStepShort"
                    }
                },
                "tags": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "worker_id": {
                    "type": "string"
                }
            }
        },
        "model.BuildStatus": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3
            ],
            "x-enum-varnames": [
                "BuildRunning",
                "BuildSuccessful",
                "BuildFailed",
                "BuildCanceled"
            ]
        },
        "model.BuildStep": {
            "type": "object",
            "properties": {
                "build_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "duration": {
                    "$ref": "#/definitions/time.Duration"
                },
                "id": {
                    "type": "string"
                },
                "logs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.BuildLog"
                    }
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.BuildStepCore": {
            "type": "object",
            "properties": {
                "build_id": {
                    "type": "string"
                },
                "duration": {
                    "$ref": "#/definitions/time.Duration"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.BuildStepShort": {
            "type": "object",
            "properties": {
                "build_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "duration": {
                    "$ref": "#/definitions/time.Duration"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "time.Duration": {
            "type": "integer",
            "enum": [
                1,
                1000,
                1000000,
                1000000000
            ],
            "x-enum-varnames": [
                "Nanosecond",
                "Microsecond",
                "Millisecond",
                "Second"
            ]
        },
        "util.Message": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}